requirements.txt
Django==1.7.5
Pillow==2.7.0
dj-database-url==0.3.0
dj-static==0.0.6
django-activity-stream==0.5.1
django-filter==0.9.2
django-localflavor==1.1
django-toolbelt==0.0.1
django-wpadmin==1.7.0
djangorestframework==3.0.5
gunicorn==19.2.1
psycopg2==2.6
static3==0.5.1
wsgiref==0.1.2
